backend/requirements.txt
# Flask==3.1.0
# Flask-Cors==5.0.0
# Flask-SQLAlchemy==3.1.1
pymysql==1.1.1
cryptography==44.0.0
pillow==11.0.0
pillow-heif==0.21.0
openai
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langsmith==0.1.62
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-langchain
llama-index-embeddings-openai
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-anyscale
llama-index-llms-langchain
llama-index-llms-openai
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
fastapi
uvicorn

llama-parse==0.4.3
llamaindex-py-client==0.1.19
pytrends
ace-tools-open
python-dotenv
selenium
webdriver_manager

mysql-connector-python
pydantic[email]

PyJWT